+++ web/CHANGES.md
@@ -2,3 +2,4 @@
 
 ## Add
 - implement `Responder` trait for `serde_json::Value`.
+- re-export `http_ws::{ResponseSender, ResponseWeakSender}` types in `xitca_web::handler::websocket` module